refactor(web): extract ChannelEditModal with labelled sections
