+++ web/src/routes/Channels.svelte
@@ -1,17 +1,14 @@
 <script>
   import { onMount } from 'svelte';
-  import { Button, Input, Select, Toggle, AlertDialog } from '@chrissnell/chonky-ui';
+  import { Button, AlertDialog } from '@chrissnell/chonky-ui';
   import { api, ApiError } from '../lib/api.js';
   import { toasts } from '../lib/stores.js';
   import { Platform } from '../lib/platform.js';
   import PageHeader from '../components/PageHeader.svelte';
-  import Modal from '../components/Modal.svelte';
-  import FormField from '../components/FormField.svelte';
   import { channelsStore, start as startChannelsStore, invalidate as refreshChannels } from '../lib/stores/channels.svelte.js';
   import { groupReferrers, totalReferrers } from '../lib/channelReferrers.js';
-  import { blankForm, rowToForm, formToPayload, validateForm } from '../lib/channelForm.js';
   import ChannelRow from './channels/ChannelRow.svelte';
-  import AndroidPttFields from './channels/AndroidPttFields.svelte';
+  import ChannelEditModal from './channels/ChannelEditModal.svelte';
 
   // The Channels page itself hydrates the shared store: this page is
   // the cheapest place for a first-visit operator to land, so it
@@ -57,63 +54,10 @@
   let putPendingId = $state(null);
   let putServerError = $state('');
   let putInFlight = $state(false);
+  // Captured tx/ptt context from the 409 save attempt; re-used on force retry.
+  let putPendingContext = $state(null);
   let putGroups = $derived(groupReferrers(putReferrers));
   let putTotal = $derived(totalReferrers(putReferrers));
-  // channel_type is a UI-only enum that drives the segmented picker
-  // (D11). It is NOT serialized; the wire shape is still
-  // input_device_id (nullable). 'modem' keeps all existing behavior;
-  // 'kiss-tnc' hides audio fields and sends input_device_id=null.
-  let form = $state({
-    name: '',
-    mode: 'aprs',
-    channel_type: 'modem',
-    input_device_id: '0', input_channel: '0',
-    output_device_id: '0', output_channel: '0',
-    modem_type: 'afsk', bit_rate: '1200', mark_freq: '1200', space_freq: '2200',
-    tx_delay_ms: '300', tx_tail_ms: '100', slot_ms: '100', persist: '63', full_dup: false,
-  });
-  let errors = $state({});
-
-  let isModemType = $derived(form.channel_type === 'modem');
-  let isTxEnabled = $derived(isModemType && form.output_device_id !== '0');
-
-  let inputDevices = $derived(audioDevices.filter(d => d.direction === 'input'));
-  let outputDevices = $derived(audioDevices.filter(d => d.direction === 'output'));
-  let inputDeviceOptions = $derived(inputDevices.map(d => ({ value: String(d.id), label: d.name })));
-  let outputDeviceOptions = $derived([
-    { value: '0', label: 'None (RX only)' },
-    ...outputDevices.map(d => ({ value: String(d.id), label: d.name })),
-  ]);
-
-  const modemOptions = [
-    { value: 'afsk', label: 'AFSK' },
-    { value: 'psk', label: 'PSK' },
-  ];
-
-  const channelOptions = [
-    { value: '0', label: '0 (Left/Mono)' },
-    { value: '1', label: '1 (Right)' },
-  ];
-
-  // Android PTT form state (only used when Platform.kind === 'android')
-  const PTT_METHOD_CP2102N_RTS = 1;
-  let androidPttMethod = $state(PTT_METHOD_CP2102N_RTS);
-  // Reference to the AndroidPttFields component instance for cleanup calls.
-  // Plain let: bind:this targets are imperative refs, not reactive state.
-  let androidPttComp = null;
-
-  // When the edit modal opens for an Android channel, populate
-  // androidPttMethod from the stored ptt config if present.
-  function loadAndroidPttMethod(row) {
-    if (Platform.kind !== 'android') return;
-    // The ptt config is returned on the channel object as row.ptt.
-    // If method === 'android', gpio_pin carries the method int.
-    if (row.ptt?.method === 'android' && row.ptt?.gpio_pin) {
-      androidPttMethod = row.ptt.gpio_pin;
-    } else {
-      androidPttMethod = PTT_METHOD_CP2102N_RTS;
-    }
-  }
 
   onMount(async () => {
     startChannelsStore();
@@ -139,36 +83,22 @@
 
   function openCreate() {
     editing = null;
-    const defaultInput = inputDevices.length > 0 ? String(inputDevices[0].id) : '0';
-    form = { ...blankForm(), input_device_id: defaultInput };
-    errors = {};
-    if (Platform.kind === 'android') {
-      androidPttMethod = PTT_METHOD_CP2102N_RTS;
-    }
     modalOpen = true;
   }
 
-  async function openEdit(row) {
+  function openEdit(row) {
     editing = row;
-    // Phase 2: input_device_id is nullable on the wire. Null means
-    // KISS-TNC-only; any non-null value means modem-backed. The
-    // segmented picker is read-only on edit (D11) — the "Convert…"
-    // link below the badge is the only way to flip it.
-    form = rowToForm(row, txTimings[row.id]);
-    errors = {};
-    loadAndroidPttMethod(row);
     modalOpen = true;
   }
 
-  function validate() {
-    errors = validateForm(form);
-    return Object.keys(errors).length === 0;
+  function closeModal() {
+    modalOpen = false;
   }
 
-  async function handleSave() {
-    if (!validate()) return;
-    const data = formToPayload(form);
-    await persistSave(data, { force: false });
+  // handleSave receives the payload + context built by ChannelEditModal
+  // and delegates to persistSave (PUT/POST + referrer-confirm path).
+  async function handleSave({ payload, isTxEnabled, txTiming, androidPttMethod }) {
+    await persistSave(payload, { force: false, isTxEnabled, txTiming, androidPttMethod });
   }
 
   // persistSave runs the actual PUT/POST + follow-up tx-timing save.
@@ -177,7 +107,10 @@
   // reload dance. `force` adds ?force=true to the PUT query when
   // true; backend treats that as "I know this breaks referrers,
   // proceed anyway" (Phase 1 handoff).
-  async function persistSave(data, { force }) {
+  // isTxEnabled, txTiming, androidPttMethod are passed from ChannelEditModal
+  // via handleSave; the force-retry path (confirmForcePut) re-passes the
+  // context it captured when the 409 landed.
+  async function persistSave(data, { force, isTxEnabled = false, txTiming = null, androidPttMethod = null }) {
     try {
       let channelId;
       if (editing) {
@@ -194,14 +127,10 @@
       }
 
       // Save TX timing if this is a TX-capable channel
-      if (isTxEnabled && channelId) {
+      if (isTxEnabled && channelId && txTiming) {
         const timingData = {
           channel: channelId,
-          tx_delay_ms: parseInt(form.tx_delay_ms, 10),
-          tx_tail_ms: parseInt(form.tx_tail_ms, 10),
-          slot_ms: parseInt(form.slot_ms, 10),
-          persist: parseInt(form.persist, 10),
-          full_dup: form.full_dup,
+          ...txTiming,
         };
         await api.put(`/tx-timing/${channelId}`, timingData);
       }
@@ -213,7 +142,7 @@
       // row's gpio_pin field as a method-int carrier rather than adding a
       // new schema column — the semantics are different from the desktop
       // gpio-line use but the field is otherwise unused on Android channels.
-      if (Platform.kind === 'android' && channelId) {
+      if (Platform.kind === 'android' && androidPttMethod != null && channelId) {
         await api.post('/ptt', {
           channel_id: channelId,
           method: 'android',
@@ -239,6 +168,7 @@
         putReferrers = err.body.referrers;
         putPendingPayload = data;
         putPendingId = editing.id;
+        putPendingContext = { isTxEnabled, txTiming, androidPttMethod };
         putServerError = err.body?.error || err.message || '';
         putConfirmOpen = true;
         return;
@@ -261,13 +191,15 @@
       if (editing?.id !== targetId) {
         editing = channels.find((c) => c.id === targetId) || editing;
       }
-      await persistSave(data, { force: true });
+      const ctx = putPendingContext || {};
+      await persistSave(data, { force: true, ...ctx });
     } finally {
       putInFlight = false;
       putConfirmOpen = false;
       putReferrers = [];
       putPendingPayload = null;
       putPendingId = null;
+      putPendingContext = null;
       putServerError = '';
     }
   }
@@ -280,6 +212,7 @@
     putReferrers = [];
     putPendingPayload = null;
     putPendingId = null;
+    putPendingContext = null;
     putServerError = '';
   }
 
@@ -395,139 +328,15 @@
   </div>
 {/if}
 
-<!-- Add/Edit modal -->
-<div class="wide-modal">
-<Modal bind:open={modalOpen} title={editing ? 'Edit Channel' : 'New Channel'}>
-  <!-- Channel type picker (D11). Segmented on create; read-only badge
-       on edit. -->
-  <div class="channel-type-row">
-    <span class="channel-type-label" id="channel-type-label">Channel type</span>
-    {#if editing}
-      <span class="channel-type-badge">
-        {#if form.channel_type === 'modem'}Modem-backed{:else}KISS-TNC only{/if}
-      </span>
-    {:else}
-      <div class="segmented" role="radiogroup" aria-labelledby="channel-type-label">
-        <button type="button"
-                role="radio"
-                aria-checked={form.channel_type === 'modem'}
-                class="segment"
-                class:active={form.channel_type === 'modem'}
-                onclick={() => form.channel_type = 'modem'}>
-          Modem-backed
-        </button>
-        <button type="button"
-                role="radio"
-                aria-checked={form.channel_type === 'kiss-tnc'}
-                class="segment"
-                class:active={form.channel_type === 'kiss-tnc'}
-                onclick={() => form.channel_type = 'kiss-tnc'}>
-          KISS-TNC only
-        </button>
-      </div>
-    {/if}
-  </div>
-
-  <div class="form-grid-2">
-    <FormField label="Name" error={errors.name} id="ch-name">
-      <Input id="ch-name" bind:value={form.name} placeholder="VHF APRS" />
-    </FormField>
-    {#if isModemType}
-      <FormField label="Modem Type" id="ch-modem">
-        <Select id="ch-modem" bind:value={form.modem_type} options={modemOptions} />
-      </FormField>
-    {/if}
-  </div>
-
-  {#if Platform.kind === 'android' && isModemType}
-    <AndroidPttFields bind:this={androidPttComp} bind:method={androidPttMethod} channelId={editing?.id} />
-  {/if}
-
-  <FormField
-    label="Mode"
-    hint="APRS only: beacon, digipeater, iGate, and messages may transmit. Packet only: AX.25 connected-mode terminal sessions only; APRS subsystems are blocked. APRS + Packet: both, on a shared channel."
-    id="ch-mode"
-  >
-    <Select
-      id="ch-mode"
-      bind:value={form.mode}
-      aria-label="Channel mode"
-      options={[
-        { value: 'aprs', label: 'APRS only' },
-        { value: 'packet', label: 'Packet only' },
-        { value: 'aprs+packet', label: 'APRS + Packet' },
-      ]}
-    />
-  </FormField>
-
-  {#if isModemType}
-    <div class="form-grid-4">
-      <FormField label="Input Device" error={errors.input_device_id} id="ch-indev">
-        <Select id="ch-indev" bind:value={form.input_device_id} options={inputDeviceOptions} />
-      </FormField>
-      <FormField label="Input Channel" id="ch-inch">
-        <Select id="ch-inch" bind:value={form.input_channel} options={channelOptions} />
-      </FormField>
-      <FormField label="Output Device" id="ch-outdev">
-        <Select id="ch-outdev" bind:value={form.output_device_id} options={outputDeviceOptions} />
-      </FormField>
-      {#if isTxEnabled}
-        <FormField label="Output Channel" id="ch-outch">
-          <Select id="ch-outch" bind:value={form.output_channel} options={channelOptions} />
-        </FormField>
-      {/if}
-    </div>
-    <div class="form-grid-3">
-      <FormField label="Bit Rate" id="ch-baud">
-        <Input id="ch-baud" bind:value={form.bit_rate} type="number" placeholder="1200" />
-      </FormField>
-      <FormField label="Mark Freq (Hz)" id="ch-mark">
-        <Input id="ch-mark" bind:value={form.mark_freq} type="number" placeholder="1200" />
-      </FormField>
-      <FormField label="Space Freq (Hz)" id="ch-space">
-        <Input id="ch-space" bind:value={form.space_freq} type="number" placeholder="2200" />
-      </FormField>
-    </div>
-
-    {#if isTxEnabled}
-      <div class="tx-timing-section">
-        <h4 class="section-label">Transmit Timing</h4>
-        <div class="form-grid-4">
-          <FormField label="TX Delay (ms)" id="ch-txd"
-            hint="Key-up time before sending. 300ms typical.">
-            <Input id="ch-txd" bind:value={form.tx_delay_ms} type="number" placeholder="300" />
-          </FormField>
-          <FormField label="TX Tail (ms)" id="ch-txt"
-            hint="Hold time after last byte. 100ms typical.">
-            <Input id="ch-txt" bind:value={form.tx_tail_ms} type="number" placeholder="100" />
-          </FormField>
-          <FormField label="Slot Time (ms)" id="ch-slot"
-            hint="CSMA listen interval. 100ms is standard.">
-            <Input id="ch-slot" bind:value={form.slot_ms} type="number" placeholder="100" />
-          </FormField>
-          <FormField label="Persistence (0-255)" id="ch-persist" error={errors.persist}
-            hint="TX probability = (val+1)/256. 63 ≈ 25%.">
-            <Input id="ch-persist" bind:value={form.persist} type="number" placeholder="63" />
-          </FormField>
-        </div>
-        <Toggle bind:checked={form.full_dup} label="Full Duplex" />
-      </div>
-    {/if}
-  {:else}
-    <div class="kiss-only-explainer">
-      This channel is serviced by a KISS TNC interface (configured on
-      the <a href="#/kiss">KISS page</a>). No audio device, modem, or
-      CSMA timing is required — frames route through the attached
-      KISS-TNC backend.
-    </div>
-  {/if}
-
-  <div class="modal-actions">
-    <Button onclick={() => { androidPttComp?.clearPttHold(); androidPttComp?.stopUsbPoll(); modalOpen = false; }}>Cancel</Button>
-    <Button variant="primary" onclick={handleSave}>{editing ? 'Save' : 'Create'}</Button>
-  </div>
-</Modal>
-</div>
+<!-- Add/Edit modal (extracted to ChannelEditModal) -->
+<ChannelEditModal
+  bind:open={modalOpen}
+  {editing}
+  {audioDevices}
+  {txTimings}
+  onSave={handleSave}
+  onCancel={closeModal}
+/>
 
 <!-- Phase 5 two-step delete: stage 1 = impact dialog (only when the
      channel has referrers). Lists what the cascade will do to each
@@ -671,47 +480,6 @@
     gap: 12px;
   }
 
-  /* Wider modal for channel editor */
-  .wide-modal :global(.modal) {
-    width: min(860px, 94vw);
-  }
-  .wide-modal :global(.modal-body) {
-    overflow-y: auto;
-  }
-  .form-grid-2 {
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    gap: 0 16px;
-  }
-  .form-grid-3 {
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    gap: 0 16px;
-  }
-  .form-grid-4 {
-    display: grid;
-    grid-template-columns: repeat(4, 1fr);
-    gap: 0 16px;
-  }
-
-  /* TX Timing section in modal */
-  .tx-timing-section {
-    margin-top: 12px;
-    padding-top: 12px;
-    border-top: 1px solid var(--border-color);
-  }
-  .section-label {
-    margin: 0 0 6px 0;
-    font-size: 15px;
-    font-weight: 600;
-  }
-
-  .modal-actions {
-    display: flex;
-    gap: 8px;
-    justify-content: flex-end;
-    margin-top: 16px;
-  }
   .modal-footer {
     display: flex;
     gap: 8px;
@@ -721,71 +489,6 @@
   :global(.danger-action) {
     background: var(--color-danger) !important;
     color: white !important;
-  }
-
-  /* D11 channel-type segmented control + edit-time read-only badge. */
-  .channel-type-row {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-    margin-bottom: 12px;
-    flex-wrap: wrap;
-  }
-  .channel-type-label {
-    font-size: 13px;
-    font-weight: 600;
-    color: var(--text-secondary);
-    min-width: 110px;
-  }
-  .segmented {
-    display: inline-flex;
-    border: 1px solid var(--border-color);
-    border-radius: var(--radius);
-    overflow: hidden;
-  }
-  .segment {
-    padding: 8px 14px;
-    min-height: 40px;
-    background: var(--bg-secondary);
-    border: none;
-    border-right: 1px solid var(--border-color);
-    color: var(--text-primary);
-    font: inherit;
-    cursor: pointer;
-  }
-  .segment:last-child {
-    border-right: none;
-  }
-  .segment.active {
-    background: var(--color-info-muted, rgba(56, 139, 253, 0.15));
-    color: var(--color-info, #388bfd);
-    font-weight: 600;
-  }
-  .segment:focus-visible {
-    outline: 2px solid var(--color-info, #388bfd);
-    outline-offset: -2px;
-  }
-
-  .channel-type-badge {
-    display: inline-block;
-    padding: 4px 10px;
-    border-radius: var(--radius);
-    background: var(--bg-tertiary);
-    font-size: 13px;
-    font-weight: 600;
-    color: var(--text-primary);
-  }
-  .kiss-only-explainer {
-    padding: 10px 12px;
-    background: var(--bg-tertiary);
-    border-left: 3px solid var(--color-info, #388bfd);
-    border-radius: var(--radius);
-    font-size: 13px;
-    color: var(--text-secondary);
-    margin-bottom: 8px;
-  }
-  .kiss-only-explainer a {
-    color: var(--color-info, #388bfd);
   }
 
   /* Phase 5 two-step delete flow */
